streamlit_exp/requirements.txt
streamlit>=1.31
gspread>=6.0.0
google-auth>=2.23.0
pandas>=2.2.0